+++ lib/commands/impl/install/install_get.dart
@@ -1,6 +1,8 @@
 import '../../../common/utils/pubspec/pubspec_utils.dart';
 
 Future<void> installGet([bool runPubGet = false]) async {
+  if (PubspecUtilsExt.useSmGetx) return;
+
   PubspecUtils.removeDependencies('get', logger: false);
   // TODO: 当 get 5 正式发布之后修改
   final isVersion5 = PubspecUtilsExt.getxVersion == 5;